Move upgrade logic into package
