+++ collect_app/src/main/java/org/odk/collect/android/application/initialization/ExistingProjectMigrator.kt
@@ -1,11 +1,12 @@
-package org.odk.collect.android.projects
+package org.odk.collect.android.application.initialization
 
 import android.content.Context
 import androidx.preference.PreferenceManager
 import org.apache.commons.io.FileUtils
-import org.odk.collect.android.application.initialization.Upgrade
+import org.odk.collect.android.application.initialization.upgrade.Upgrade
 import org.odk.collect.android.preferences.keys.MetaKeys
 import org.odk.collect.android.preferences.source.SettingsProvider
+import org.odk.collect.android.projects.CurrentProjectProvider
 import org.odk.collect.android.storage.StoragePathProvider
 import org.odk.collect.projects.Project
 import org.odk.collect.projects.ProjectsRepository
