+++ collect_app/src/main/java/org/odk/collect/android/application/initialization/upgrade/AppUpgrader.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.application.initialization
+package org.odk.collect.android.application.initialization.upgrade
 
 import org.odk.collect.shared.Settings
